reorganized the classes into separate packages
