+++ src/main/java/co/grandcircus/TutorApp2019/controller/HomeController.java
@@ -18,17 +18,16 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.client.RestTemplate;
 import org.springframework.web.servlet.ModelAndView;
 
-import co.grandcircus.TutorApp2019.entity.Business;
-import co.grandcircus.TutorApp2019.entity.BusinessMarks;
-import co.grandcircus.TutorApp2019.entity.BusinessResults;
-import co.grandcircus.TutorApp2019.entity.Coordinates;
-import co.grandcircus.TutorApp2019.entity.GoogleMarks;
-import co.grandcircus.TutorApp2019.entity.MapData;
-import co.grandcircus.TutorApp2019.entity.PlaceResults;
 import co.grandcircus.TutorApp2019.entity.Student;
 import co.grandcircus.TutorApp2019.entity.TimeLedger;
 import co.grandcircus.TutorApp2019.entity.Tutor;
-import co.grandcircus.TutorApp2019.entity.TutorMarks;
+import co.grandcircus.TutorApp2019.json.Business;
+import co.grandcircus.TutorApp2019.json.BusinessResults;
+import co.grandcircus.TutorApp2019.json.Coordinates;
+import co.grandcircus.TutorApp2019.json.MapData;
+import co.grandcircus.TutorApp2019.json.PlaceResults;
+import co.grandcircus.TutorApp2019.marks.BusinessMarks;
+import co.grandcircus.TutorApp2019.marks.TutorMarks;
 import co.grandcircus.TutorApp2019.repo.StudentRepo;
 import co.grandcircus.TutorApp2019.repo.TimeLedgerRepo;
 import co.grandcircus.TutorApp2019.repo.TutorRepo;
