+++ src/main/java/co/grandcircus/TutorApp2019/json/Geometry.java
@@ -1,4 +1,4 @@
-package co.grandcircus.TutorApp2019.entity;
+package co.grandcircus.TutorApp2019.json;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
