+++ src/main/java/co/grandcircus/TutorApp2019/marks/GoogleMarks.java
@@ -1,4 +1,4 @@
-package co.grandcircus.TutorApp2019.entity;
+package co.grandcircus.TutorApp2019.marks;
 
 public abstract class GoogleMarks {
